+++ cwt/_cwt.py
@@ -810,7 +810,7 @@ def compute_cwt_paper_hd(activity, scales, wavelet_f0, step_slug):
     # iwave_test = wavelet.icwt(coefs, scales, delta_t, wavelet=wavelet_type)
     # plt.plot(iwave_test)
     # plt.show()
-    return coefs.astype(np.complex64), coi.astype(np.float16), scales, freqs, wavelet_type, delta_t
+    return coefs.astype(np.complex64), coi.astype(float), scales, freqs, wavelet_type, delta_t
 
 
 def compute_cwt_new(y):
@@ -955,7 +955,7 @@ def dwt_power(
             levels,
         )
     # dwt_data = np.hstack(coeffs)
-    dwt_data = np.hstack(cc).astype(np.float16)
+    dwt_data = np.hstack(cc).astype(float)
     return dwt_data, dwt_data.shape, 1, dwt_data.shape[0], cc.real, cc.shape, wavelet, mode
 
 
@@ -1088,9 +1088,9 @@ def cwt_power(
         power_masked = power_cwt.copy()
 
     # return coefs.copy().real, freqs, coi, power_masked.shape, scales
-    imag = coefs.copy().imag.astype(np.float16)
-    real = coefs.copy().real.astype(np.float16)
-    mag = abs(coefs.copy()).astype(np.float16)
+    imag = coefs.copy().imag.astype(float)
+    real = coefs.copy().real.astype(float)
+    mag = abs(coefs.copy()).astype(float)
     if enable_coi:
         imag = mask_cwt(imag.copy(), coi)
         real = mask_cwt(real.copy(), coi)
